change that the search component is responsible for all the logic and the suggestions component is responsible for presenting the suggestions
