+++ client/src/components/Results/searchResults.module.css
@@ -4,6 +4,8 @@
     margin: 0 auto;
     padding: 20px;
     background-color: #ffe6f2;
+    border-radius: 10px;
+
   }
 
 .resultsList {
