+++ client/src/components/Search/searchComponent.js
@@ -1,12 +1,14 @@
 import React, { useState } from 'react';
 import axios from 'axios';
-import Autocomplete from '../AutoComplete/autocompleteComponent';
+import Suggestions from '../Suggestions/suggestionsComponent';
 import SearchResults from '../Results/searchResultsComponent';
 
 const SearchComponent = () => {
   const [query, setQuery] = useState(''); // saves the string that the user entered
   const [results, setResults] = useState([]); // saves the search results received from the server after the search
   const [searchPerformed, setSearchPerformed] = useState(false); // saves if the search was performed
+  const [suggestions, setSuggestions] = useState([]); // saves the suggestions
+
 
   const handleSearch = async () => {
     try {
@@ -19,6 +21,23 @@ const SearchComponent = () => {
     }
   };
 
+  const handleSuggestions = async (e) => {
+    const value = e.target.value
+    setQuery(value);
+
+    if (value.length === 0) {
+        // reset suggestions when query is empty
+        setSuggestions([]);
+      } else if(value.length > 0) {
+        try {
+            const response = await axios.post('http://localhost:5000/api/content/search/suggestions', { query: value });
+            setSuggestions(response.data);
+        } catch (error) {
+            console.error('Error fetching suggestions:', error);
+        };
+      }
+    };
+
   const handleKeyDown  = (event) => {
     if (event.key === 'Enter') {
       handleSearch();
@@ -27,7 +46,14 @@ const SearchComponent = () => {
 
   return (
     <div>
-      <Autocomplete setQuery={setQuery} onKeyDown={handleKeyDown} />
+      <input
+        type="text"
+        placeholder="Search for content..."
+        value={query}
+        onChange={handleSuggestions}
+        onKeyDown={handleKeyDown}
+      />
+      <Suggestions setQuery={setQuery} suggestions={suggestions} setSuggestions={setSuggestions} />
       <button onClick={handleSearch}>Search </button>
       <SearchResults results={results} searchPerformed={searchPerformed}  />
     </div>
