+++ client/src/components/Suggestions/suggestionsComponent.module.css
@@ -16,7 +16,8 @@
     padding-left: 0;
     margin: 0; 
     text-align: center;
- 
+  
+    
   }
   
   /*.suggestionsList*/ li {
